+++ tasks/e2e.sh
@@ -18,8 +18,8 @@ function cleanup {
   echo 'Cleaning up.'
   cd $root_path
   # Uncomment when snapshot testing is enabled by default:
-  # rm ./template/src/__snapshots__/App.test.js.snap
-  rm -rf $temp_cli_path $temp_app_path $clean_path
+  # rm ./packages/react-scripts/template/src/__snapshots__/App.test.js.snap
+  rm -rf $temp_cli_path $temp_app_path
 }
 
 # Error messages are redirected to stderr
@@ -36,6 +36,10 @@ function handle_exit {
   exit
 }
 
+function create_react_app {
+  node "$temp_cli_path"/node_modules/create-react-app/index.js $*
+}
+
 # Exit the script with a helpful error message when any error is encountered
 trap 'set +x; handle_error $LINENO $BASH_COMMAND' ERR
 
@@ -49,15 +53,15 @@ set -x
 cd ..
 root_path=$PWD
 
-# Lint
+npm install
+
+# Lint own code
 ./node_modules/.bin/eslint --ignore-path .gitignore ./
 
 # ******************************************************************************
 # First, test the create-react-app development environment.
 # This does not affect our users but makes sure we can develop it.
 # ******************************************************************************
-
-npm install
 
 # Test local build command
 npm run build
@@ -80,36 +84,19 @@ npm start -- --smoke-test
 # Next, pack react-scripts and create-react-app so we can verify they work.
 # ******************************************************************************
 
-# Pack CLI (it doesn't need cleaning)
-cd global-cli
-npm install
+# Pack CLI
+cd $root_path/packages/create-react-app
 cli_path=$PWD/`npm pack`
 
-# Packing react-scripts takes more work because we want to clean it up first.
-# Create a temporary clean folder that contains production only code.
-# Do not overwrite any files in the current folder.
-clean_path=`mktemp -d 2>/dev/null || mktemp -d -t 'clean_path'`
+# Go to react-scripts
+cd $root_path/packages/react-scripts
 
-# Copy some of the project files to the temporary folder.
-# Exclude folders that definitely won’t be part of the package from processing.
-# We will strip the dev-only code there, `npm pack`, and copy the package back.
-cd $root_path
-rsync -av --exclude='.git' --exclude=$clean_path\
-  --exclude='node_modules' --exclude='build'\
-  './' $clean_path  >/dev/null
-
-# Open the clean folder
-cd $clean_path
-# Now remove all the code relevant to development of Create React App.
-files="$(find -L . -name "*.js" -type f)"
-for file in $files; do
-  sed -i.bak '/\/\/ @remove-on-publish-begin/,/\/\/ @remove-on-publish-end/d' $file
-  rm $file.bak
-done
+# Like bundle-deps, this script modifies packages/react-scripts/package.json,
+# copying own dependencies (those in the `packages` dir) to bundledDependencies
+node $root_path/tasks/bundle-own-deps.js
 
 # Finally, pack react-scripts
-npm install
-scripts_path=$clean_path/`npm pack`
+scripts_path=$root_path/packages/react-scripts/`npm pack`
 
 # ******************************************************************************
 # Now that we have packed them, create a clean app folder and install them.
@@ -124,7 +111,7 @@ npm install $cli_path
 # Install the app in a temporary location
 temp_app_path=`mktemp -d 2>/dev/null || mktemp -d -t 'temp_app_path'`
 cd $temp_app_path
-node "$temp_cli_path"/node_modules/create-react-app/index.js --scripts-version=$scripts_path test-app
+create_react_app --scripts-version=$scripts_path test-app
 
 # ******************************************************************************
 # Now that we used create-react-app to create an app depending on react-scripts,
@@ -155,8 +142,14 @@ npm start -- --smoke-test
 # Finally, let's check that everything still works after ejecting.
 # ******************************************************************************
 
-# Eject
+# Eject...
 echo yes | npm run eject
+
+# ...but still link to the local packages
+npm link $root_path/packages/babel-preset-react-app
+npm link $root_path/packages/eslint-config-react-app
+npm link $root_path/packages/react-dev-utils
+npm link $root_path/packages/react-scripts
 
 # Test the build
 npm run build
@@ -177,6 +170,42 @@ npm test -- --watch=no
 
 # Test the server
 npm start -- --smoke-test
+
+
+# ******************************************************************************
+# Test --scripts-version with a version number
+# ******************************************************************************
+
+cd $temp_app_path
+create_react_app --scripts-version=0.4.0 test-app-version-number
+cd test-app-version-number
+
+# Check corresponding scripts version is installed.
+test -e node_modules/react-scripts
+grep '"version": "0.4.0"' node_modules/react-scripts/package.json
+
+# ******************************************************************************
+# Test --scripts-version with a tarball url
+# ******************************************************************************
+
+cd $temp_app_path
+create_react_app --scripts-version=https://registry.npmjs.org/react-scripts/-/react-scripts-0.4.0.tgz test-app-tarball-url
+cd test-app-tarball-url
+
+# Check corresponding scripts version is installed.
+test -e node_modules/react-scripts
+grep '"version": "0.4.0"' node_modules/react-scripts/package.json
+
+# ******************************************************************************
+# Test --scripts-version with a custom fork of react-scripts
+# ******************************************************************************
+
+cd $temp_app_path
+create_react_app --scripts-version=react-scripts-fork test-app-fork
+cd test-app-fork
+
+# Check corresponding scripts version is installed.
+test -e node_modules/react-scripts-fork
 
 # Cleanup
 cleanup